fix(providers/gh): correctly pass callback to vim.system
The continuation callback argument to vim.system is the third one.
Without this patch, the following error is raised since the callback is
passed as the first argument:

> cmd: expected table, got function

This was missed in the migration to the new async conventions
in fdc5b4fc3f5300c87fbde16cab0aeb122ba6324a.